Cleaning.

+++ include/espressif/spi_flash.h
@@ -6,6 +6,9 @@
 #ifndef __SPI_FLASH_H__
 #define __SPI_FLASH_H__
 
+#include <stddef.h>
+#include <stdint.h>
+
 typedef enum {
     SPI_FLASH_RESULT_OK,
     SPI_FLASH_RESULT_ERR,
@@ -14,8 +17,8 @@ typedef enum {
 
 #define SPI_FLASH_SEC_SIZE      4096
 
-SpiFlashOpResult spi_flash_erase_sector(uint16 sec);
-SpiFlashOpResult spi_flash_write(uint32 des_addr, uint32 *src_addr, uint32 size);
-SpiFlashOpResult spi_flash_read(uint32 src_addr, uint32 *des_addr, uint32 size);
+SpiFlashOpResult spi_flash_erase_sector(int sec);
+SpiFlashOpResult spi_flash_write(uint32_t des_addr, uint32_t *src_addr, size_t size);
+SpiFlashOpResult spi_flash_read(uint32_t src_addr, uint32_t *des_addr, size_t size);
 
 #endif
